partly refactor to remove a bug
