+++ react/src/app-components/home-components/selection-components/restaurant-information-components/ButtonToInvite.js
@@ -7,9 +7,8 @@ import ListItem from '@material-ui/core/ListItem'
 import ListItemText from '@material-ui/core/ListItemText';
 import DialogTitle from '@material-ui/core/DialogTitle'
 import { makeStyles,withStyles } from '@material-ui/core/styles'
-import { CopyToClipboard } from 'react-copy-to-clipboard'
 // 他ファイルからインポート
-import InviteIcon from './../../../icon/InviteIcon'
+import InviteIcon from './../../../../icon/InviteIcon'
 import InviteModal from './InviteModal'
 /*
 招待ボタン：押すと招待URLが表示される
